Extract math operation from dataset.h
