+++ dataset/CMakeLists.txt
@@ -6,6 +6,7 @@ set(INC_FILES
     include/scipp/dataset/dataset.h
     include/scipp/dataset/dataset_index.h
     include/scipp/dataset/except.h
+    include/scipp/dataset/math.h
     include/scipp/dataset/reduction.h
     include/scipp/dataset/shape.h
     include/scipp/dataset/sort.h
@@ -22,6 +23,7 @@ set(SRC_FILES
     groupby.cpp
     histogram.cpp
     map_view.cpp
+    math.cpp
     operations.cpp
     reduction.cpp
     shape.cpp
