+++ dataset/include/scipp/dataset/dataset.h
@@ -812,9 +812,6 @@ SCIPP_DATASET_EXPORT DataArray rebin(const DataArrayConstView &a, const Dim dim,
 SCIPP_DATASET_EXPORT Dataset rebin(const DatasetConstView &d, const Dim dim,
                                    const VariableConstView &coord);
 
-[[nodiscard]] SCIPP_DATASET_EXPORT DataArray
-reciprocal(const DataArrayConstView &a);
-
 /// Return one of the inputs if they are the same, throw otherwise.
 template <class T> T same(const T &a, const T &b) {
   core::expect::equals(a, b);
