+++ dataset/include/scipp/dataset/math.h
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2020 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#pragma once
+
+#include "scipp/dataset/dataset.h"
+
+namespace scipp::dataset {
+
+[[nodiscard]] SCIPP_DATASET_EXPORT DataArray
+reciprocal(const DataArrayConstView &a);
+
+} // namespace scipp::dataset
